retro and blogging: post logic update
