+++ README.md
@@ -1,36 +1,36 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+# Portfolio Website
+
+A personal portfolio website built with Next.js featuring a retro Windows 98 theme.
 
 ## Getting Started
 
-First, run the development server:
+Install dependencies and run the development server:
 
 ```bash
+npm install
 npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
 ```
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+Open [http://localhost:3000](http://localhost:3000) to view the site.
 
-You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
+## Theme
 
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
+This project uses [98.css](https://github.com/jdan/98.css) - a CSS framework for building faithful recreations of old UIs. The site features a Windows 98-inspired design with dark mode support.
 
-## Learn More
+## Documentation
 
-To learn more about Next.js, take a look at the following resources:
+For more information, see the [`docs/`](./docs/) directory:
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+- [Adding Blog Posts](./docs/ADDING_BLOG_POSTS.md) - How to add new blog posts
+- [Decap CMS Plan](./docs/DECAP_CMS_PLAN.md) - Future CMS integration roadmap
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+## Tech Stack
 
-## Deploy on Vercel
+- **Framework:** [Next.js](https://nextjs.org/) with App Router
+- **Styling:** [98.css](https://github.com/jdan/98.css) + Tailwind CSS
+- **Content:** Markdown files in `src/content/blog/`
+- **Language:** TypeScript
 
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
+## Deploy
 
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+Deploy on [Vercel](https://vercel.com) or your preferred platform. The site builds as a standard Next.js application.
