+++ package.json
@@ -14,6 +14,8 @@
     "98.css": "^0.1.21",
     "dompurify": "^3.3.1",
     "framer-motion": "^12.29.0",
+    "gray-matter": "^4.0.3",
+    "marked": "^17.0.1",
     "next": "^16.1.4",
     "node": "^22.10.0",
     "react": "^19.2.3",
@@ -22,6 +24,7 @@
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
+    "@types/marked": "^5.0.2",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
