+++ src/content/blog/hello-world.md
@@ -1,0 +1,18 @@
+---
+title: "Hello World - My First Blog Post"
+date: "2026-02-01"
+excerpt: "Welcome to my blog! This is my first post where I share my thoughts on web development and life."
+---
+
+Welcome to my blog! This is my first post where I share my thoughts on web development and life.
+
+I've been working on this portfolio website and decided to give it a retro Windows 98 theme. It's been a fun project combining modern web technologies with nostalgic design elements.
+
+In this blog, I'll be sharing:
+
+- My experiences as a developer
+- Projects I'm working on
+- Thoughts on technology and coding
+- Random musings about life
+
+Stay tuned for more posts!
